# weft/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Weft API
  version: 0.2.0
  description: |
    Coordinator service for Loominal multi-agent infrastructure.

    Weft routes work, manages agent lifecycle, handles scaling, and provides a REST API
    for fleet management. It runs as a multi-tenant service that handles multiple projects
    in a single deployment.

    ## Key Features
    - **Agent Management**: Register, discover, and shutdown agents
    - **Work Distribution**: Submit and track work items across agents
    - **Spin-Up Targets**: Manage dynamic agent provisioning
    - **Batch Operations**: Perform bulk operations on agents, targets, and work
    - **Pagination**: All list endpoints support cursor-based pagination
    - **Semantic Context**: Responses include resolved agent details for better decision-making

  contact:
    name: Michael LoPresti
    email: mike@lopresti.org

  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://weft.loominal.dev
    description: Production server
...
paths:
...
  /api/targets:
    get:
      summary: List spin-up targets
      description: |
        Returns a paginated list of registered spin-up targets with optional filtering.
        Includes semantic context with last spin-up information and agent details.
      operationId: listTargets
      tags:
        - Targets
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          description: Filter by agent type
          schema:
            $ref: '#/components/schemas/AgentType'
        - name: status
          in: query
          description: Filter by target status
          schema:
            $ref: '#/components/schemas/TargetStatus'
        - name: capability
          in: query
          description: Filter by capability
          schema:
            type: string
        - name: boundary
          in: query
          description: Filter by allowed boundary
          schema:
            type: string
        - name: classification
          in: query
          deprecated: true
          description: (Deprecated) Use 'boundary' instead
          schema:
            type: string
        - name: cursor
          in: query
          description: Pagination cursor
          schema:
            type: string
        - name: limit
          in: query
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: List of targets with pagination
          headers:
            X-Deprecated-Param:
              description: Indicates use of deprecated parameter
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required:
                  - targets
                  - count
                  - hasMore
                properties:
                  targets:
                    type: array
                    items:
                      $ref: '#/components/schemas/TargetResponse'
                  count:
                    type: integer
                  total:
                    type: integer
                  hasMore:
                    type: boolean
                  nextCursor:
                    type: string
                    nullable: true
                  prevCursor:
                    type: string
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Register new spin-up target
      description: |
        Registers a new target for dynamic agent provisioning.
        Targets can use various mechanisms (SSH, Kubernetes, GitHub Actions, etc.)
      operationId: registerTarget
      tags:
        - Targets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetRegisterRequest'
            examples:
              ssh:
                summary: SSH target
                value:
                  name: home-server
                  description: Home lab server
                  agentType: claude-code
                  capabilities: [typescript, python]
                  boundaries: [personal]
                  mechanism: ssh
                  config:
                    mechanism: ssh
                    ssh:
                      host: server.home.lan
                      user: agent
                      command: /usr/local/bin/claude-code
              kubernetes:
                summary: Kubernetes target
                value:
                  name: k8s-claude
                  agentType: claude-code
                  capabilities: [typescript]
                  mechanism: kubernetes
                  config:
                    mechanism: kubernetes
                    kubernetes:
                      namespace: loominal
                      jobNamePrefix: claude-agent
                      image: ghcr.io/loominal/claude-code:latest
      responses:
        '201':
          description: Target registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/targets/{id}:
    get:
      summary: Get target details
      description: Returns detailed information about a specific target including last spin-up event
      operationId: getTarget
      tags:
        - Targets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Target ID or name
          schema:
            type: string
      responses:
        '200':
          description: Target details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update target
      description: Updates an existing spin-up target configuration
      operationId: updateTarget
      tags:
        - Targets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Target ID or name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                capabilities:
                  type: array
                  items:
                    type: string
                boundaries:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Target updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Remove target
      description: Removes a spin-up target from the registry
      operationId: removeTarget
      tags:
        - Targets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Target ID or name
          schema:
            type: string
      responses:
        '200':
          description: Target removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/targets/{id}/test:
    post:
      summary: Health check target
      description: Performs a health check on a spin-up target to verify it's reachable
      operationId: testTarget
      tags:
        - Targets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Target ID or name
          schema:
            type: string
      responses:
        '200':
          description: Health check results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/targets/{id}/spin-up:
    post:
      summary: Trigger spin-up for target
      description: Manually triggers agent provisioning using the specified target
      operationId: spinUpTarget
      tags:
        - Targets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Target ID or name
          schema:
            type: string
      responses:
        '200':
          description: Spin-up initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpinUpResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/targets/{id}/disable:
    post:
      summary: Disable target
      description: Disables a target to prevent future spin-ups
      operationId: disableTarget
      tags:
        - Targets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Target disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/targets/{id}/enable:
    post:
      summary: Enable target
      description: Re-enables a previously disabled target
      operationId: enableTarget
      tags:
        - Targets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Target enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/targets/disable-batch:
    post:
      summary: Batch disable targets
      description: Disables multiple targets matching filter criteria or specific IDs
      operationId: batchDisableTargets
      tags:
        - Targets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDisableTargetsRequest'
            example:
              filter:
                status: error
                mechanism: ssh
              preventSpinUp: true
      responses:
        '200':
          description: Batch operation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDisableTargetsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication (configure via API_TOKENS env var)
  schemas:
    AgentType:
      type: string
      enum:
        - copilot-cli
        - claude-code
      description: Type of AI agent
...
    TargetStatus:
      type: string
      enum:
        - available
        - in-use
        - disabled
        - error
      description: Status of a spin-up target
...
    AgentSummary:
      type: object
      description: Summary information about an agent for semantic context
      required:
        - guid
        - agentType
      properties:
        guid:
          type: string
          format: uuid
          description: Unique agent identifier
        handle:
          type: string
          description: Agent handle/username
        agentType:
          $ref: '#/components/schemas/AgentType'
        hostname:
          type: string
          description: Hostname where agent is running
...
    SpinUpEvent:
      type: object
      description: Information about a spin-up event
      required:
        - time
        - outcome
      properties:
        time:
          type: string
          format: date-time
        agent:
          $ref: '#/components/schemas/AgentSummary'
        workItemId:
          type: string
        outcome:
          type: string
          enum: [success, failure]
        error:
          type: string
    TargetResponse:
      type: object
      description: Spin-up target with semantic context (last spin-up details)
      required:
        - id
        - name
        - agentType
        - capabilities
        - boundaries
        - mechanism
        - status
        - healthStatus
        - registeredBy
        - registeredAt
        - updatedAt
        - useCount
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        agentType:
          $ref: '#/components/schemas/AgentType'
        capabilities:
          type: array
          items:
            type: string
        boundaries:
          type: array
          items:
            type: string
        mechanism:
          type: string
          enum: [ssh, github-actions, local, webhook, kubernetes]
        config:
          type: object
          description: Mechanism-specific configuration
        status:
          $ref: '#/components/schemas/TargetStatus'
        currentAgentGuid:
          type: string
          format: uuid
        healthStatus:
          type: string
          enum: [healthy, unhealthy, unknown]
        lastError:
          type: string
        registeredBy:
          type: string
          enum: [cli, api, agent, config]
        registeredByAgentGuid:
          type: string
          format: uuid
        registeredAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
        useCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        lastSpinUp:
          $ref: '#/components/schemas/SpinUpEvent'
          description: Information about the last spin-up event
    TargetRegisterRequest:
      type: object
      required:
        - name
        - agentType
        - capabilities
        - mechanism
        - config
      properties:
        name:
          type: string
        description:
          type: string
        agentType:
          $ref: '#/components/schemas/AgentType'
        capabilities:
          type: array
          items:
            type: string
        boundaries:
          type: array
          items:
            type: string
        mechanism:
          type: string
          enum: [ssh, github-actions, local, webhook, kubernetes]
        config:
          type: object
          description: Mechanism-specific configuration
        healthCheck:
          type: object
          properties:
            enabled:
              type: boolean
            intervalMs:
              type: integer
            timeoutMs:
              type: integer
        tags:
          type: array
          items:
            type: string
    SpinUpResult:
      type: object
      required:
        - success
        - targetId
        - targetName
        - timestamp
      properties:
        success:
          type: boolean
        targetId:
          type: string
        targetName:
          type: string
        error:
          type: string
        mechanismResult:
          type: object
          properties:
            pid:
              type: integer
            runId:
              type: integer
            jobName:
              type: string
            response:
              type: object
        timestamp:
          type: string
          format: date-time
    HealthCheckResult:
      type: object
      required:
        - targetId
        - healthy
        - timestamp
      properties:
        targetId:
          type: string
        healthy:
          type: boolean
        error:
          type: string
        responseTimeMs:
          type: integer
        timestamp:
          type: string
          format: date-time
...
    BatchDisableTargetsRequest:
      type: object
      properties:
        filter:
          type: object
          properties:
            agentType:
              $ref: '#/components/schemas/AgentType'
            status:
              type: string
            mechanism:
              type: string
        targetIds:
          type: array
          items:
            type: string
        preventSpinUp:
          type: boolean
          default: true
    BatchDisableTargetsResponse:
      type: object
      required:
        - success
        - failed
        - count
        - errors
        - completedAt
        - totalProcessed
        - successRate
        - disabledTargets
        - alreadyDisabled
      properties:
        success:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            type: string
        count:
          type: integer
        errors:
          type: object
          additionalProperties:
            type: string
        completedAt:
          type: string
          format: date-time
        totalProcessed:
          type: integer
        successRate:
          type: number
        disabledTargets:
          type: array
          items:
            type: string
        alreadyDisabled:
          type: array
          items:
            type: string
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: ValidationError
            message: Invalid request parameters
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or missing authentication token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: InternalServerError
            message: An unexpected error occurred